apps/ingestors/file-downloader/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "file_downloader" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=90 --cov-report html:'../../../coverage/apps/ingestors/file-downloader/html' --cov-report xml:'../../../coverage/apps/ingestors/file-downloader/coverage.xml' --html='../../../reports/apps/ingestors/file-downloader/unittests/html/index.html' --junitxml='../../../reports/apps/ingestors/file-downloader/unittests/junit.xml'"

[tool.poetry]
name = "file_downloader"
version = "1.0.0"
description = "File"
authors = [ ]
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "file_downloader"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.11"
  requests = "^2.31.0"
  warlock = "^2.0.1"

    [tool.poetry.dependencies.pylog]
    path = "../../../libs/python/pylog"
    develop = true

    [tool.poetry.dependencies.core]
    path = "../../../libs/python/core"
    develop = true

    [tool.poetry.dependencies.pyrabbitmq]
    path = "../../../libs/python/pyrabbitmq"
    develop = true

    [tool.poetry.dependencies.pysd]
    path = "../../../libs/python/pysd"
    develop = true

    [tool.poetry.dependencies.py-minio]
    path = "../../../libs/storage-client/python/py-minio"
    develop = true

    [tool.poetry.dependencies.config-loader]
    path = "../../../libs/service-setup/python/config-loader"
    develop = true

    [tool.poetry.dependencies.py-repository]
    path = "../../../libs/api-client/python/py-repository"
    develop = true

    [tool.poetry.dependencies.dto-events]
    path = "../../../libs/dtos/python/dto-events"
    develop = true

    [tool.poetry.dependencies.dto-gateway]
    path = "../../../libs/dtos/python/dto-gateway"
    develop = true

    [tool.poetry.dependencies.dto-serializer]
    path = "../../../libs/dtos/python/serializer"
    develop = true

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
